Refactor: Re-organize packages
